+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Services/RoleAssignmentsMapper.cs
@@ -67,8 +67,6 @@ using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using CoreApiModel = RadixDlt.CoreApiSdk.Model;
 using GatewayModel = RadixDlt.NetworkGateway.GatewayApiSdk.Model;